Fix bug in putting previous scheduled send time in compose toolbar
If the user adds the send time text box to the compose toolbar and
then edits a previously scheduled message, the previous send time is
supposed to show up in the text box automatically, but instead the
word "long" was showing up because I forgot a couple of curly braces
in a date formatting string. I'm amazed that this has been broken for
almost six months and no one has mentioned it to me before now.