Makefile fix: add "all" and "clean" rules, with "all" the default
The first rule in a Makefile is what gets executed when "make" is run
with no arguments. By convention this is supposed to be a rule called
"all". Add an "all" rule at the top that builds "send_later.xpi".

Also add a "clean" rule (another one that should be present by
convention) which deletes "send_later.xpi".