Makefile: Rebuild send_later.xpi if dev/include-manifest changes
If dev/include-manifest changes, then that could mean that a file has
been removed from the manifest, or that there is a new file listed in
the manifest whose timestamp is older than the last time
send_later.xpi was rebuilt. Therefore we need to rebuild
send_later.xpi if it is older than dev/include-manifest.