Provide a message interface for other add-ons to get or send Send Later options
This restores the ability that exists in TB68 for other add-ons to
update Send Later's preferences as needed.

Arguably Send Later should pop up an alert when this happens asking
the user if they want to grant permission for the other add-on to do
what it's trying to do, but this is at least a good initial
implementation.