Prefer the future when parsing a date/time entered by hand
When the user is using the text box to schedule a message rather than
the date and time pickers, prefer future times when what is entered is
ambiguous. This way, for example, if the user enters "10:00" and it's
later than that in the day, the next day is chosen instead of a time
in the past. This restores the behavior of the TB68 version of Send
Later.