When disabling the add-on, perform all actions sequentially
When disabling the add-on because it failed to initialize properly,
instead of just firing off a bunch of promises which will execute in
random order based on whenever the JS engine decides to fulfill each
one, execute them all sequentially. This eliminates a bunch of
unnecessary nondeterminism which could cause intermittent issues we
would need to debug, and it also means we will know reliably that the
disable steps, some of which may depend on earlier steps, happen in
order.