Remove unnecessary jiggering of compose action which causes issues
After converting the compose action click handler to an async
function, the first time the compose action button is clicked it
doesn't work, and an error is logged that the popup can only be opened
from a user input handler, which is weird because it _is_ being opened
from a user input handler.

Having said that, simplifying the code by not unnecessarily setting
and clearing the popup file makes this problem go away, and this
change doesn't hurt anything else, so let's go with that.