Fix two bugs in the Send Later column in the message list
1) I don't understand why there are two custom column handlers in the
experiments directory, but it appears at some point the column
handling was switched from customColumn* to legacyColumn*, which
changed the argument types for
`messenger.columnHandler.setColumnVisible`, but the code in
`background.js` wasn't updated to reflect that change. Now it has
been, so the Send Later column is now properly enabled and disabled.

2) If the Drafts folder is the currently displayed folder when
Thunderbird starts up, the Send Later column is now properly enabled
and made visible if it's supposed to be.