Make Ctrl-Click and Shift-Click on the compose action button work
You're supposed to be able to bind shortcuts that get activated if you
click the compose action button while holding down Ctrl or Shift, but
that was broken. It's fixed now. The issue was that the click events
don't get delivered when there's a popup bound to the button, so we
have to leave the popup unbound most of the time, then when we get a
click, first check if one of the modifier keys was held down, and if
not, then add the popup to the button for long enough to open it and
trigger the open. This is as partial fix for #583.