Use event keys, not codes, to work properly with remapped keyboards
When reacting to key events, use the `key` field of the events rather
than the `code` field, because the latter refers to the physical
keyboard key and the former refers to the key as remapped by the user,
which is what we obviously want. If they've remapped their keyboard we
want to obey the remapping!

This is the last piece of the fix for #583. It may also be a fix for