Merge TB102 and TB115 support into a single extension
When I did the port to TB115 I initially decided to simplify it by
stripping out TB102 code and having two separate forks, one for
continuing to maintain TB102 support and one for TB115 moving forward.

On further reflection I think that's a bad idea, because:

* I don't want to have to keep cherry-picking TB115 bug fixes into the
  TB102 branch and fixing merge conflicts.
* I don't want to have to keep putting out two releases for every
  release.
* I don't want to make the ATN reviewers have to review two versions
  of the add-on for every release.

I am therefore merging the TB102 support back into main and supporting
TB102+ on that main until we decide to remove TB102 support entirely,
at which point all of the TB102 code will be clearly delineated and
easy to find and remove.

In the process of doing this merge-back of the TB102 functionality
into main, I also made the following changes:

* Update the header view implementation to treat different 3pane
  windows as distinct with distinct values for the header field.
* Simplify the header view API; there's no reason for us to be passing
  in an object instead of a plain argument.
* Delete obsolete type definitions from the header view schema.
* Remove an obsolete call to messenger.headerView.onHeaderRowUpdate
  that should have been removed before but was missed.