Don't try to parse scheduling pop-up inputs repeatedly
It appears that when the scheduling pop-up opens we get a ton of
change events for it. We only want the code in `onInput` to execute
once, though, so add code to cache the previously inputs we last
handled in `onInput` and only do our thing if they've changed.

This is just good hygiene, but the specific reason why I made this
change is because if someone has a dynamic scheduling function that
invokes a web API, and they make that scheduling function their
default in the pop-up, then when they open the pop-up that scheduling
function was getting called for every invocation of `onInput`,
flooding the web API with a bunch of unnecessary, redundant calls.