Work around issue causing messages not to send for Owl users
There is something very weird going on when Thunderbird is connected
to an Exchange account via Owl for Exchange. The call into the
`messages` API to delete a draft message after it is delivered is
causing our code to stop executing at the point where that call is
made. I have no idea what's going on there or how to fix it properly,
so in the meantime this commit contains a workaround which isn't
optimal but isn't too bad for me to feel like I can't ship it.