Don't store bad custom date formats in preferences
Check the custom date formats entered by the user in the preferences
and don't store invalid or incomplete ones.

Also: use an input event handler instead of a per-second interval
function to update the sample date displayed next to each custom date
format in the preferences. The interval function was an unnecessary
performance drain. Also, since it was catching all exceptions it was
continuing to run forever even after the options window was close.
Just make all this go away by using an event handler.

REQUIRES TRANSLATION. There's a new string that requires translation
in this commit.