Fix mismatch between UI locale and OS locale in Send Later pop-up
We are using the Intl API to format the date/time that gets inserted
into the text field of the pop-up window, and the SugarJS API to parse
the date in that text field. The Intl API uses the OS locale setting
to determine how to format dates, while we were using the UI locale
setting to determine how SugarJS should parse them. This means that,
e.g., if the user's OS locale was en-GB and their UI locale was en-US,
then dates would be formatted DD/MM/YYYY but SugarJS was expecting
them to be formatted MM/DD/YYYY. This is fixed! Now we use a clever
trick (ref: https://stackoverflow.com/a/57529410) to get the OS locale
from the Intl API and give that to SugarJS instead of the UI locale.