New feature: scheduling message(s) from the message list
You can now select one or more messages in the message list, right
click, and select the "Schedule message(s) menu command near the
bottom of the menu to pop up a Send Later scheduling dialog and
specify when you want the selected message(s) to be sent. When you
then click the button to actually do the scheduling, Send Later opens
each of the messages one by one and schedules them automatically, then
when it's done it closes the scheduling pop-up.