Ask user to confirm context menu commands before any side-effects
Rather than displaying the pop-up asking the user to confirm send now
or put in outbox after the first message is opened, do it before.
Otherwise if we ever get to the point where we are able to edit an
existing draft rather than needing to create a new one (ugh), even if
the user cancels by then we will have already removed the scheduling
info from the draft by opening it for editing.