Put a hyphen before "Mozilla Thunderbird" in pop-up title bars
I don't know when this broke, or if it's always been broken, but at
least in TB115 on Linux, when titlePreface for a pop-up is set to just
"Foo", the actual title bar looks like "FooMozilla Thunderbird", so
let's add " - " to the end of our preface so it looks like the much
more attractive "Foo - Mozilla Thunderbird".

I don't know if the behavior I'm seeing is a bug that will be fixed at
some point or what, but at least for the time being this makes things
look better.