Improve handling of attempts to schedule with encryption / signing
Delete the old EncryptionIncompat* translation strings and replace
them with new EncryptionUnsupported* strings that say something
different and more accurate.

Check when a message is being scheduled whether it has encryption or
signing enabled, or whether encrypting drafts is enabled, and if so,
pop up a message warning the user.