Store the scheduled send time in the Date field of scheduled drafts
Because we can't currently add a custom column to the message list in
the drafts folder, it's difficult for users to see scheduled drafts.
To address this, let's put the scheduled send time of the draft into
the Date field by default. If you don't like this behavior you can
disable it by setting the avanced preference "scheduledDateField" to
false.

Fixes #598.