Don't force the Send Later toolbar button visible in compose windows
As noted in #559, we should not be forcing the Send Later button to be
visible in the compose window if the user doesn't want it to be.

Instead, if the button isn't visible when we're trying to open the
pop-up, then we can just open a standalone window instead.

This requires a bit of trickery to tell the standalone window what
compose tab it's associated with so it can do the right thing with
regard to settings that are stored in that tab.

Fixed #559.