Clean up handling of translation strings surrounding daily etc.
There were two sets of translation strings for "minutely", "daily",
etc. The redundancy has been eliminated.

I'm getting rid of the English translation "minutely", which is too
cute by half, and changing it to "every minute."

I've added translation string descriptions for the remaining "daily"
etc. strings.

I've fixed a bug which was causing the UI in the pop-up to show a
recurrence as e.g. "every 1 minutes" when it should have been saying
"every minute", or "every 1 days" instead of "daily", etc.

As part of that, I fixed the code that parses recurrence specs to
convert more things from strings to numbers, and to use a consistent
method for doing that.