Improve handling of rescheduling errors, corrupt folders, late messages
Sometimes the error pop-up about the user's drafts folder being
corrupt might be displayed when that wasn't actually the problem. I
hope this has now been fixed. Specifically, we're catching errors
about things other than folder corruption and displaying a separate
warning about those, and not repeatedly displaying the corrupt folder
warning.

The message locking functionality has been isolated in a separate
class, and it's smarter now. We store different lock reasons in it, in
particular: the generic "this message was sent" reason, which is what
leads to the corrupt Drafts folder warning if the message is seen
again; a message being past its grace period; and a failure while
trying to reschedule a message.

Furthermore, we're now removing entries in the lock cache after six
months, and to make that feasible, we're enforcing a delivery grace
period of six months max even if the user has that preference
disabled.

The pop-up which appears when a message is too late to be delivered
has been updated to reflect that fact, and how old the late message is
is now displayed in a more user-friendly format.