Clean up obsolete preferences, remove redundant hard-coded prefs list
Remove all the preferences that are no longer being used from the
default preferences list.

Remove translation strings associated with obsolete preferences.

Replace the hard-coded list of preference keys with one that is
dynamically generated from the default preferences list.

Migrate the checkTimePref_isMilliseconds preference to get rid of the
isMilliseconds thing. It's not necessary now that the checkTimePref
can be a floating-point number.