Restore a bunch of tooltips that were lost in the WebExtension conversion
When the add-on was converted from legacy to WebExtension, a bunch of
useful tooltips explaining what stuff does was lost. This commit
restores them.

In order to do this I had to switch from using the localize.js that
the add-on was previously using to automatically localize HTML files,
to [i18n.js][i18njs] provided by the Thunderbird team. My version of
their script is modified slightly to put the listener at the bottom of
the script so that it doesn't need to be put separately into the
script associated with each HTML page.

[i18njs]: https://github.com/thundernest/addon-developer-support/tree/master/scripts/i18n