Workaround for "Send Now" access key not working on macOS
The issue here is that either Firefox isn't handling "dead keys"
properly on macOS when both Ctrl and Option are pressed, or that the
macOS input layer is completely preventing Firefox from handling dead
keys properly so there's nothing it can do.

The problem is that Option-N is a dead key on macOS, because you type
Option-N and then another character to get that character with a tilde
over it. But *Ctrl*-Option-N shouldn't be interpreted as a dead key,
because the Ctrl key being held down as well should prevent that, but
it's not. I don't know whether it's macOS or Firefox that's doing the
wrong thing.

In any case this workaround looks at the key code to find out if
that's happening, which won't work if someone has remapped the key on
their keyboard, but will do what the user expects in the vast majority
of cases.

Note: I've included some telemetry in this commit so I can find out
how often this is happening because if it's happening a lot that will
give me ammunition to convince the Firefox folks to try to do
something about it, if they can (i.e., if this isn't macOS screwing
them over).

Ref: https://bugzilla.mozilla.org/show_bug.cgi?id=1850454

Fixes #462.