Use Google Translate and extended-thunder.github.io for privacy policy
We want to send people to a localized version of the privacy policy,
so we'll use Google Translate for that like we do for the release
notes, but we can't do that on addons.thunderbird.net because we need
to fix the links that are broken by Google Translate, and while we can
add script code on our site to do that, we can't do that on
addons.thunderbird.net. So let's move the privacy policy over to our
site.