Go back to requiring the compose window button to be visible
Unfortunately openPopup in TB102 doesn't return a status indicating
whether it was successful, so we can't tell whether the popup opened
or not, so we can't tell whether we need to open a separate popup.
Therefore for tb102 we need to go back to making forcing the toolbar
button to be visible.