Actually use both browser.alarms and setTimeout for main loop
I thought I was using both browser.alarms and setTimeout for the main
loop, but it turns out when I implemented that functionality I
commented out the setTimeout part and then forgot to uncomment it
before committing, so we were just using browser.alarms.

Which would be fine if it weren't for the fact that the documentation
claims you can't have more than one alarm per minute. That doesn't
actually seem to be true in practice, but let's assume that it is and
make sure the code is robust against that.

It actually wasn't good enough to just uncomment the commented-out
line, because there was a bug in my code which was sometimes causing
both the setTimeout timeout and the browser.alarms alarm to trigger
for the same cycle, which we obviously don't want. So this commit also
refactors the code to ensure that only one of them will trigger.