Fix bug when rescheduling a message which can't be sent today
When there are day-of-week restrictions on a message and its overdue
to be sent but can't be sent today because of the day-of-week
restrictions, we attempt to reschedule it for a day when it can be
sent. However, there was a bug in that code so it was erroring out. As
a result we would spam the error console with an exception every time
through the main loop until we reached the day when it was OK to send
the message. This has been fixed.