Make return receipts work and don't allow scheduling with DSN enabled
Make return receipts work for messages scheduled with Send Later.

If the user tries to schedule a message with Send Later when the
Delivery Status Notification option is enabled, Send Later blocks the
scheduling because that option cannot be supported by Send Later. The
user can choose to either turn off the option or send the message
without Send Later.

Fixes #509.