Add whitelist for `Send delays messages' and `Send does Send Later'
You can now specify the name of an address book in your preferences,
and the address book exists and all of the recipients of a message are
listed in it, then Send Later is bypassed on send even if you've
enabled Send delays messages or Send does Send Later.

Fixes #264.