If all else fails, use ISO format for dates in the popup text field
SugarJS can't parse everything. If we can't parse the date format in
the OS or the one the user has specified, then we shouldn't be
inserting it into the text field in the scheduling popup when copying
the value from the date and time pickers into the text field.
Therefore, when formatting that value we now make sure we can parse
it, and if not, we use ISO format, which is guaranteed to be parseable
by Date.parse. Not great, but sufficient to make it work for everyone.

The problem prompting this change was reported by someone using
zh_Hant_TW as their locale, but it's also an issue for zh and zh_TW.
SugarjS doesn't know how to handle the former at all, and its ability
to parse the latter isn't great.