Add internal logging
Send Later now has the ability to keep an internal log in local
storage that persists across Thunderbird sessions. This is primarily
useful for debugging issues that take a long time to manifest and/or
manifest at an unpredictable time so they aren't discovered right
away, because in either of those cases Thunderbird may get restarted
before the issue is discovered in which case the logs related to it
are gone from the error console.

This adds a new "View internal log" link to the main window pop-up
which opens the internal log page. Internal logging is disabled by
default. It can be enabled, disabled, or cleared on the internal log
page, and there's a "Copy" button on the internal log page which
allows the contents of the log to be copied to the clipboard so they
can be easily pasted into a file or email message.