Handle the message ID changing when moving a draft to a new folder
My previous assumption that the ID of a message changed when it was
moved from one folder to another was incorrect.

Nevertheless, my code that was marking the draft as read using its old
message ID was working! I have no idea why, but that's clearly a
fluke, so we shouldn't rely on that.

Therefore, we now mark the message as read _before_ we move it to the
new folder, while we still have its ID.

Also, to make sure we don't make the mistake in the future of trying
to use the old message ID after moving the message, I'm now setting it
to null after the move.

Also, I've updated the code to throw an exception if saving the draft
fails without Thunderbird throwing an exception of its own, which
should never happen, but let's be careful.

Also, I've updated the code to clarify that saving the draft only ever
returns 1 or 2 messages, either just the draft or the draft plus the
additional FCC set by the user while composing the message.

Also, we are no longer marking the FCC as read, we're just marking the
actual draft.