TB bug: save draft twice when editing from subfolder
See https://bugzilla.mozilla.org/show_bug.cgi?id=1855487 and the
comment included in this commit.

In some cases if a user goes to edit a message in a draft subfolder,
Send Later doesn't successfully remove the scheduling information from
the message, so if the scheduled send time comes while the user is
editing it, or if the user never saves the message after starting to
edit it, it could get sent at the previously scheduled time despite
the fact that the user didn't intend that.