Alert at delivery time if a message is too large
Due to the Thunderbird bug referenced in #643, we can't deliver
message through the outbox if they are larger than
mailnews.message_warning_size. The best we can do is alert the user.
This isn't great but it's better than all Send Later message delivery
getting clogged.

Fixes #643 probably as well as we can under the circumstances.