Don't move scheduled drafts into subfolder if already there
Thunderbird was changed in release 115.2 so that if you edit a draft
that's in a subfolder of your Drafts folder and then save it, it goes
back into the same folder; previously it was saved into the default
Drafts folder for the account, not the subfolder. This broke Send
Later's "Store scheduled messages in subfolder" feature: if you edited
a scheduled message in a subfolder and then scheduled the edited
message, it would be saved in a subfolder of the subfolder instead of
saving in the subfolder it started in. This is now fixed.