Restore old values in pop-up when editing previously scheduled message
At some point the code for restoring the old values to the scheduling
pop-up when editing a previously scheduled message broke. It is now
fixed.

Saving the values entered into the pop-up while editing a message,
such that if the user closes and reopens the pop-up their changes are
still there, is now working again as well.

Fixes #702.