Yet another attempt to fix the mailbox corruption race condition
Move draft compacting back to being a synchronous operation.

Instead of compacting shortly after a message is scheduled or
rescheduled, wait until the beginning of the next main loop run.

Fix a bug which was preventing compacted folders from being removed
from the list of folders to be compacted, so they were being compacted
repeatedly unnecessarily.