Another attempt to fix folder corruption: check folder mod times
Using the activity manager to detect when Thunderbird was modifying
the Drafts folders didn't work. Therefore, this commit undoes that
functionality and attempts yet another new idea for avoiding
corruption: check the modification times on the folder files and wait
until they've been unmodified for a full second before compacting.