Add a preference controlling whether to detach the scheduling popup
Add a preference controlling whether the scheduling popup should be
detached from the compose window. Right now, the default value of this
preference is true, i.e., the previous behavior is preserved, because
[Thunderbird bug 1916112][bug1916112] means that at least for some
people the popup window will be blank when it is attached to the
compose window, so we're not ready to restore that behavior yet.

[bug1916112]: https://bugzilla.mozilla.org/show_bug.cgi?id=1916112