Prevent concurrent expunging / compacting of Drafts folders
If `compactDrafts` was waiting for all the folders to become idle and
`compactDrafts` was called again during the waiting, then both would
end up running in parallel and attempting to expunge / compact all the
Drafts folders. Fix this by making subsequent invocations wait for
prior invocations to complete, and by localizing the list of folders
to clean at the start of the loop other code can add new folders to
clean without disrupting the current clean cycle.