Add telemetry event for evaluating a dynamic scheduling function
The way Send Later is currently evaluating dynamic scheduling
functions is no longer permitted, and it's going to take a substantial
amount of work to change how it's done to make it permissible. I need
to figure out whether it's worth the effort or whether, instead, the
feature is so infrequently used that we should just remove the feature
from the add-on. This new telemetry event—which logs just the fact
that a dynamic scheduling function was executed, not what code was
executed or what the inputs or the outputs were—will enable us to get
a rough estimate of how many people are using this feature to drive
our decision about whether to do the work to continue supporting it.