TB139 compatibility: mail.identity.default.draft_folder pref gone
We can no longer attempt to determine the default drafts folder name
from the preference `mail.identity.default.draft_folder`, because that
preference no longer exists. This was never a necessary thing to do,
strictly speaking; it was more of a "throw everything at the wall and
see what sticks" kind of thing, so we're just removing it rather than
trying to replace it with something compatible with TB139.

Fixes #752.