Try harder to resize popups to make contents visible
Sometimes when we try to make the contents of a popup visible,
Thunderbird hasn't finished rendering the new content so the body
height is wrong so we don't make it big enough. To work around this,
try 10 times, 10ms apart, to do the resize. With empirical testing
I've never seen it take more than 50ms so I'm doubling that to be
safe.