Fix exceptionHandling when doNextRecur fails
The `catch` block for handling exceptions in `doNextRecur`, the
function that schedules the next recurring instance of a recurring
message after the current instance was just sent by Send Later, had
two bugs in it that caused the exception-handling block to itself
generate an exception. Given that `doNextRecur` isn't actually
supposed to generate exceptions under normal operating conditions,
it's not clear that this ever would have been an issue for anyone in
the real world; still, we should handle exceptions more gracefully, so
let's clean this up.